% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{missing_pcs}
\alias{missing_pcs}
\title{\emph{Missing} Postcode Sectors (\code{PCS})}
\format{
A list including two data.table, \code{PCS} and \code{PCD}, with the following columns for \code{PCS}:
\describe{
\item{\code{PCS}}{ The missing Postcode Sector }
\item{\code{PCS.map}}{ The Postcode Sector included at its place }
}
}
\usage{
missing_pcs
}
\description{
The Postcode Sectors in the provided boundaries are built using Census 2021 Output Areas as a basis.
Unfortunately, mostly because the two system are built some output area contains more thanone postcode sector.
This table lists the sectors that are missed from the boundaries,
and the corresponding postcode Sector that is
}
\keyword{datasets}
