% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{output_areas}
\alias{output_areas}
\title{Output Areas}
\format{
A data.table with the following columns:
\describe{
\item{\code{OA}}{ Output Area ONS code: Census 2021 for England and Wales, Census 2011 for N.Ireland and Scotland}
\item{\code{PCS}}{ Postcode Sector }
\item{\code{PCD}}{ Postcode District }
\item{\code{PCT}}{ Post Town }
\item{\code{PCA}}{ Postcode Area }
\item{\code{RGN}}{ ONS code for Region (England only; the other Regions assumed the following pseudo codes:
NIE_RGN = Northern Ireland, SCO_RGN = Scotland, WLS_RGN = Wales) }
\item{\code{RGNd}}{ Description for Region }
\item{\code{CTRY}}{ Country 3-chars code: ENG = England, NIE = Northern Ireland, SCO = Scotland, WLS = Wales }
}

For further details, see the \href{https://geoportal.statistics.gov.uk/search?collection=Dataset&sort=name&tags=all(PRD_ONSPD)}{ONS Geoportal}
}
\usage{
output_areas
}
\description{
A list of all 2,619,057 \emph{Postcode Units} \code{PCU} in the UK (as of NOV-22: 1,737,670 active, 881,387 terminated),
together with their geographic coordinates (CRS 4326, WGS84), and the corresponding \emph{Output Area} \code{OA} and current \emph{Postcode Sector} \code{PCS}.
}
\keyword{datasets}
