% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uprn.R
\name{pc2uprn}
\alias{pc2uprn}
\title{Regional UPRN file name for postal hierarchy}
\usage{
pc2uprn(x, sf = TRUE)
}
\arguments{
\item{x}{a string equal to a Postcode unit or a zone in the \emph{Postal} hierarchy}

\item{sf}{logical to indicate if the output should be a spatial \code{sf} object instead of a \code{data.table}}
}
\value{
a \code{data.table} or an \code{sf} object
}
\description{
Given a Postcode Unit or a code for a Zone in the Postal Hierarchy (see either the table \code{pzones} or \code{postal}),
returns the subset of all the locations of the Properties in that Zone.
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
