% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pca_totals}
\alias{pca_totals}
\title{Summaries by Postcode Areas (\code{PCA})}
\format{
A list including two data.table, \code{PCS} and \code{PCD}, with the following columns for \code{PCS}:
\describe{
\item{\code{PCS.old}}{ The actual Postcode Sector corresponding to the rule }
\item{\code{PCS}}{ The current Postcode Sector corresponding to its location }
}
and similarly for \code{PCD}.
}
\usage{
pca_totals
}
\description{
A mapping between Postcode Sectors/Districts, related to terminated Units (but each table contains all records).
}
\keyword{datasets}
