% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PCD}
\alias{PCD}
\title{PCD}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 2799 rows and 2 columns.
}
\usage{
PCD
}
\value{
A \code{sf} dataframe with only one \code{PCD} column for the corresponding \emph{Postcode District} codes.
}
\description{
Digital Vector Boundaries in \code{sf} format and \emph{WGS84} CRS (\emph{EPSG} 4326) for the \emph{Postcode Districts} in the UK.
}
\details{
Built by dissolving the \code{PCS} boundaries using the \code{lookups} table.
}
\note{
These are \emph{not} the official boundaries as released by Royal Mail with their \href{https://www.poweredbypaf.com/}{Postcode Address File (PAF)},
but only an approximation using the \emph{Output Areas}
and the \href{https://geoportal.statistics.gov.uk/search?collection=Dataset&sort=-created&tags=all(PRD_ONSPD)}{ONS Postcode Directory} as a basis.
}
\keyword{datasets}
