% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pcd_linkage}
\alias{pcd_linkage}
\title{Postcode Districts Linkage}
\format{
A data.table with the following two columns:
\describe{
\item{\code{PCD.old}}{ The actual Postcode District corresponding to the rule }
\item{\code{PCD}}{ The current Postcode District corresponding to its location }
}
}
\usage{
pcd_linkage
}
\description{
A mapping between Postcode Districts corresponding to the rule of \emph{four chars},
and the geographical position of the Postcode Units that currently form(ed) them.
}
\keyword{datasets}
