% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pcs_linkage}
\alias{pcs_linkage}
\title{Postcode Sectors Linkage}
\format{
A data.table with the following two columns:
\describe{
\item{\code{PCS.old}}{ The actual Postcode Sector corresponding to the rule }
\item{\code{PCS}}{ The current Postcode Sector corresponding to its location }
}
}
\usage{
pcs_linkage
}
\description{
A mapping between Postcode Sectors corresponding to the rule of \emph{five chars},
and the geographical position of the Postcode Units that currently form(ed) them.
}
\keyword{datasets}
