% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pcs_non_geo}
\alias{pcs_non_geo}
\title{\emph{Non-Geographics} Postcode Sectors (\code{PCS})}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 1636 rows and 1 columns.
}
\usage{
pcs_non_geo
}
\description{
A list of Postcode Sectors linked to: postboxes, processing and mail sorting centres, bulk deliveries, ...
}
\keyword{datasets}
