% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pcs_regions}
\alias{pcs_regions}
\title{Postcode Sectors overlapping Regions}
\format{
A data.table with the following two columns:
\describe{
\item{\code{PCS}}{ The Postcode Sector }
\item{\code{RGN}}{ The ONS code for the Region }
}
}
\usage{
pcs_regions
}
\description{
A list of Postcode Sectors that overlap different Regions, with the Regions involved.
Notice that some of them even overlap Countries.
}
\keyword{datasets}
