% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcu_clean.R
\name{pcu_clean}
\alias{pcu_clean}
\title{Check a character string is a UK valid postcode, then convert it to a 7-char format}
\usage{
pcu_clean(x)
}
\arguments{
\item{x}{a string}
}
\value{
a 7-char string
}
\description{
The UK postcode system is hierarchical, the top level being "Postcode Area" (PCA) identified by 1 or 2 alphabetical character.
The next level is the "Postcode District" (PCD), also commonly known as the "outcode", and can take on several different formats, and anywhere from 2 to 4 alphanumeric characters long.
Next comes the Postcode Sector" (PCS), always identified by a single number, then finally the "unit", always formed by two alphabetical characters.
The combination of "sector" and "unit" is often called "incode", which is always 1 numeric character followed by 2 alphabetical characters.
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
