% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcu_clean.R
\name{pcu_coords}
\alias{pcu_coords}
\title{Extract the coordinates of a postcode}
\usage{
pcu_coords(x)
}
\arguments{
\item{x}{a PostCode Unit in any valid form}
}
\value{
A character vector of two components: longitude and latitude
}
\description{
Extract the coordinates of a postcode
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
