% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcu_map.R
\name{pcu_map}
\alias{pcu_map}
\title{Map a postcode unit}
\usage{
pcu_map(x, bfr = 10, cnc = 2)
}
\arguments{
\item{x}{a string}

\item{bfr}{the value in meter for the buffer around the UPRNs}

\item{cnc}{the value for the \emph{concavity} of the "concave hull" around the UPRNs}
}
\value{
a leaflet object
}
\description{
Build a leaflet map for a postcode using its ONS centroid, the buffer from the union of its UPRNs,
and the \emph{voronoi} polygon
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
