% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcu_clean.R
\name{pcus_clean}
\alias{pcus_clean}
\title{Clean Multiple strings for UK postcodes}
\usage{
pcus_clean(x)
}
\arguments{
\item{x}{a character vector}
}
\value{
a character vector
}
\description{
Check multiple character strings are valid UK postcodes, converting all of them to a 7-char format
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
