% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{postal}
\alias{postal}
\title{Postal Zones Lookups}
\format{
A data.table with the following columns:
\describe{
\item{\code{PCS}}{ Postcode Sector }
\item{\code{PCD}}{ Postcode District }
\item{\code{PCT}}{ Post Town }
\item{\code{PCA}}{ Postcode Area }
\item{\code{RGN}}{ ONS code for Region (England only; the other Regions assumed the following pseudo codes:
NIE_RGN = Northern Ireland, SCO_RGN = Scotland, WLS_RGN = Wales) }
}
}
\usage{
postal
}
\description{
A hierarchical lookups between Postal Zones, from \code{PCS} to \code{PCD} to \code{PCT} and \code{PCA}.
}
\keyword{datasets}
