% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{postcodes}
\alias{postcodes}
\title{Postcodes}
\format{
A data.table with the following columns:
\describe{
\item{\code{PCU}}{ Postcode Unit in 7-chars format (the column \code{PCD} in the \code{ONSPD} file):
- 2, 3 or 4 character \emph{outward code}, left aligned;
- 3 character \emph{inward code}, right aligned;
- 3rd and 4th characters may be blank }
\item{\code{is_active}}{ Flag that indicates if the corresponding unit postcode is currently active or terminated }
\item{\code{usertype}}{ Shows whether the postcode is a small or large user: 0 = small; 1 = large }
\item{\code{x_lon}}{ The longitude coordinate of the geometric centroid }
\item{\code{y_lat}}{ The latitude coordinate of the geometric centroid }
\item{\code{OA}}{ Output Area ONS code as of Census 2021 (currently England and Wales Only)}
\item{\code{OA11}}{ Output Area ONS code as of Census 2011 ()}
\item{\code{PCS}}{ Postcode Sector }
\item{\code{RGN}}{ Region ONS code (for England only; the other Regions assumed the following pseudo codes:
NIE_RGN = Northern Ireland, SCO_RGN = Scotland, WLS_RGN = Wales) }
\item{\code{CTRY}}{ Country 3-chars code: ENG = England, NIE = Northern Ireland, SCO = Scotland, WLS = Wales }
\item{\code{WPZ}}{ Workplace Zone Census 2011 }
}
}
\usage{
postcodes
}
\description{
A list of all 2,619,057 \emph{Postcode Units} \code{PCU} in the UK (as of NOV-22: 1,737,670 active, 881,387 terminated),
together with their geographic coordinates (CRS 4326, WGS84), and the corresponding \emph{Output Area} \code{OA} and current \emph{Postcode Sector} \code{PCS}.
}
\note{
The postcodes units included are only the ones with an associated grid reference, and not related to a \emph{non-geographical} Postcode Sector.

For further details, see the \href{https://geoportal.statistics.gov.uk/search?collection=Dataset&sort=name&tags=all(PRD_ONSPD)}{ONS Geoportal}
}
\keyword{datasets}
