% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pzones}
\alias{pzones}
\title{A list of codes and names for all zones related to the Postal hierarchy.}
\format{
A data.table with the following columns:
\describe{
\item{\code{type}}{ }
\item{\code{code}}{  }
\item{\code{name}}{  }
\item{\code{ordering}}{  }
}
}
\usage{
pzones
}
\description{
A list of codes and names for all zones related to the Postal hierarchy.
}
\keyword{datasets}
