% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RpostcodesUK.R
\docType{package}
\name{RpostcodesUK}
\alias{RpostcodesUK}
\title{RpostcodesUK}
\description{
Data and functionalities about the UK Postal Hierarchy:
Postcode Units, Sectors, Districts, and Areas, plus Post Towns.
}
