% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{U1}
\alias{U1}
\title{U1}
\format{
A data.table with the following columns:
\describe{
\item{\code{ PCU }}{ The Postcode Unit the Property is included into }
\item{\code{ x_lon }}{ The longitude of the Property }
\item{\code{ y_lat }}{ The latitude of the Property }
\item{\code{ N }}{ The count of references for the location }
}

For further details, see the either the \href{https://geoportal.statistics.gov.uk/search?collection=Dataset&sort=-created&tags=all(PRD_NSUL)}{NSUL}
of the \href{https://geoportal.statistics.gov.uk/search?collection=Dataset&sort=-created&tags=all(PRD_ONSUD)}{ONSUD} product.
}
\usage{
U1
}
\description{
This dataset contains the locations and correspondent Postcode Units (\code{PCU}) of all \emph{Unique Property Reference Number} (\code{UPRN})
in the Region \emph{North East} (ONS code: \code{E12000001})
}
\keyword{datasets}
